fix: migrate improvements
